Move some namespaces on bundle command
