+++ phpstan.neon
@@ -10,6 +10,7 @@ parameters:
     reportUnmatchedIgnoredErrors: false
 
     excludePaths:
+        - %currentWorkingDirectory%/src/Bundle/Command/*
         - %currentWorkingDirectory%/src/Bundle/Controller/*
         - %currentWorkingDirectory%/src/Bundle/DependencyInjection/Configuration.php
         - %currentWorkingDirectory%/src/Bundle/DependencyInjection/PagerfantaConfiguration.php
